Replace keyword filtering with BM25 relevance scoring
Use rank_bm25 to score jobs against query terms from keywords.txt instead
of binary substring matching. Scores are min-max normalized to 0-1 with
configurable --min-relevance threshold and --top-k budget cap (default 100).
Write jobs_relevance.jsonl for observability before filtering.

Also: rename profile.txt to profile.md, change LLM score scale from
0-100 int to 0-1 float (display as 0-100 only in HTML report).

+++ job_scraper/models.py
@@ -34,7 +34,7 @@ class ScoredJob:
     description: str
     source: str
     scraped_at: str
-    score: int
+    score: float
     why: str
     team: str | None = None
     posted: str | None = None
@@ -50,5 +50,5 @@ class ScoredJob:
         return cls(**{k: v for k, v in d.items() if k in valid})
 
 
-def scored_job(job: Job, score: int, why: str) -> ScoredJob:
+def scored_job(job: Job, score: float, why: str) -> ScoredJob:
     return ScoredJob(**job.to_dict(), score=score, why=why)
